+++ docs/context-reporter.md
@@ -1,4 +1,4 @@
-# Reporting Context
+# Context Reporter: Test Tags and Integration with TestRail
 
 ## Overview
