+++ src/support/index.ts
@@ -8,4 +8,4 @@ export * from './jsErrorsLogger';
 export * from './jfaker';
 export * from './browserHelper';
 export * from './modSince';
-export * from './reportingContext';
+export * from './contextReporter';
