+++ src/support/registerSupport.ts
@@ -7,7 +7,7 @@ import {repeatUntil} from './repeatUntil';
 import {step} from './testStep';
 import {jfaker} from './jfaker';
 import {modSince} from './modSince';
-import {collect as contextCollector} from './reportingContext';
+import {collect as contextCollector} from './contextReporter';
 
 export const registerSupport = (): void => {
     Cypress.Commands.add('apolloClient', apolloClient);
